fixed some syntax error

+++ protected/controller/CloudController.php
@@ -57,7 +57,7 @@ class CloudController extends BaseController
         $db=new Model('disk_file');
         $condition=array('uid'=>$_SESSION['uid'],'deleted'=>0, 'path'=>$path,'filename'=>$filename);
         $result=$db->find($condition);
-        if(empty($result))ERR::Catcher(6003);
+        if($result==false)ERR::Catcher(6003);
         //文件
         if($result['is_dir']==0)
         {
@@ -78,7 +78,7 @@ class CloudController extends BaseController
         //改变空间大小
         $condition=array('path like :path and uid=:uid and deleted=0',':path'=>arg('path').'%',':uid'=>$_SESSION['uid']);
         $result=$db->findAll($condition);
-        if(empty($result))ERR::Catcher(1002);
+        if($result==false)ERR::Catcher(1002);
         $sumsize=0;
         for($i=0;$i<count($result);$i++){
             $sumsize=$sumsize+$result[$i]['filesize'];
@@ -108,10 +108,10 @@ class CloudController extends BaseController
         if (!$this->islogin) ERR::Catcher(2001); 
         if (!arg('path')) ERR::Catcher(1003);
         if(!self::is_path_legal(arg('path'))) ERR::Catcher(1004);
+        if(!self::is_path_existed(arg('path'))) ERR::Catcher(6002);
         $name=self::getName(arg('path'));
         $filename=$name[0];
         $path=$name[1];
-        if(!self::is_path_existed($path)) ERR::Catcher(6002);
         $db=new Model("disk_file");
         $newrow=array(
             "uid"       =>$_SESSION['uid'],
@@ -143,8 +143,8 @@ class CloudController extends BaseController
         $name2=self::getName(arg('dst'));
         $filename2=$name2[0];
         $path2=$name2[1];
-        $users = new Model('users');
-        $result = $users->find(['uid=:uid', ':uid'=>$_SESSION['uid']]);
+        $db = new Model('users');
+        $result = $db->find(['uid=:uid', ':uid'=>$_SESSION['uid']]);
         $maxsize=$result['capacity']-$result['used'];
         $db=new Model("disk_file");
         $condition=array('uid'=>$_SESSION['uid'],'filename'=>$filename1,'path'=>$path1);
@@ -197,16 +197,16 @@ class CloudController extends BaseController
         $path=$name[1];
         $extension=strrchr($filename, '.');
         $condition=array('uid'=>$_SESSION['uid'],'deleted'=>0, 'path'=>$path,'filename'=>$filename);
+        $db=new Model("disk_file");
         $result=$db->find($condition);
-        if(empty($result))ERR::Catcher(6003);
-        $hash=$result['hash']; 
+        if($result==false)ERR::Catcher(6003);
+        $hash=$result['hash'];
         $dir=CONFIG::GET('CLOUD_FILE_DIRECTORY').DS.substr($hash, 0, 2).DS.substr($hash, 2);
         $mime=array('.txt'=>'text/plain','.jpg'=>'image/jpeg','.pdf'=>'application/pdf','.bmp'=>'image/x-windows-bmp','.png'=>'image/png');
         if(!array_key_exists($extension,$mime))ERROR::Catcher(1005);
         header('Content-Type:'.$mime[$extension]);
         header("filename:".$filename);
         readfile($dir);
-        SUCCESS::Catcher('success',TRUE);
         exit;       
     }
 
@@ -221,16 +221,17 @@ class CloudController extends BaseController
         $filename=$name[0];
         $path=$name[1];
         $condition=array('uid'=>$_SESSION['uid'],'deleted'=>0, 'path'=>$path,'filename'=>$filename);
+        $db=new Model("disk_file");
         $result=$db->find($condition);
-        if(empty($result))ERR::Catcher(6003);
+        if($result==false)ERR::Catcher(6003);
         $hash=$result['hash']; 
         $dir=CONFIG::GET('CLOUD_FILE_DIRECTORY').DS.substr($hash, 0, 2).DS.substr($hash, 2);         
         header ("Content-type: application/octet-stream");    
         header ("Accept-Ranges: bytes");    
         header ("Accept-Length: ".filesize($dir) );    
         header ( "Content-Disposition: attachment; filename=".$filename );       
-        readfile($file);    
-        SUCCESS::Catcher('success',TRUE);  
+        readfile($dir);
+        exit;
     }
 
     //普通上传
@@ -247,35 +248,20 @@ class CloudController extends BaseController
         $db=new Model('disk_file');
         $condition=array('uid'=>$_SESSION['uid'],'deleted'=>0, 'path'=>$path,'filename'=>$filename);
         $result=$db->find($condition);
-        if(!empty($result))ERR::Catcher(6004);
+        if(!$result==false)ERR::Catcher(6004);
         //上传文件
         $fileInfo=$_FILES['file'];
-        $users = new Model('users');
-        $result = $users->find(['uid=:uid', ':uid'=>$_SESSION['uid']]);
+        $db = new Model('users');
+        $result = $db->find(['uid=:uid', ':uid'=>$_SESSION['uid']]);
         $maxsize=$result['capacity']-$result['used'];
         $filesize=$fileInfo['size'];
-        $hash=getSHA($fileInfo);
+        $hash=self::getSHA($fileInfo);
         $dir=CONFIG::GET('CLOUD_FILE_DIRECTORY').DS.substr($hash, 0, 2).DS.substr($hash, 2);
     
         if($fileInfo['error']>0)ERR::Catcher(6005);      
-        if(is_uploaded_file($fileInfo['tmp_name']))
-        {
-            if($fileInfo['size']<=$maxsize)
-            {
-                if(!file_exists($dir))
-                {
-                    move_uploaded_file($fileInfo['tmp_name'],$dir);
-                    
-                }
-            }else
-            {
-                ERR::Catcher(6005); 
-            }
-        }else
-        {
-            ERR::Catcher(6005);
-
-        }
+        if(!is_uploaded_file($fileInfo['tmp_name']))ERR::Catcher(6005);
+        if(!$fileInfo['size']<=$maxsize)ERR::Catcher(6005);
+        if(!file_exists($dir))move_uploaded_file($fileInfo['tmp_name'],$dir);
         
         //数据库操作
         $newrow=array(
@@ -288,9 +274,9 @@ class CloudController extends BaseController
         );
         $db=new Model('disk_file');
         $db->create($newrow);
-        if($result==FALSE)ERR::Catcher(1002);
+        if($result==false)ERR::Catcher(1002);
         //剩余空间
-        $users = new Model('users');
+        $db = new Model('users');
         $condition=array('uid'=>$_SESSION['uid']);
         $result=$db->find($condition);
         $used=$result['used']+$filesize;
@@ -316,7 +302,7 @@ class CloudController extends BaseController
         $dir=CONFIG::GET('CLOUD_FILE_DIRECTORY').DS.substr($hash, 0, 2).DS.substr($hash, 2);
         
         //数据库上传
-        if(!file_exists($dir))ERR::Cacther(6003);
+        if(!file_exists($dir))ERR::Catcher(6003);
         $newrow=array(
             'uid'     =>$_SESSION['uid'],
             'path'    =>$path,
@@ -328,9 +314,9 @@ class CloudController extends BaseController
         );
         $db=new Model('disk_file');
         $db->create($newrow);
-        if($result==FALSE)ERR::Catcher(1002);
+        if($result==false)ERR::Catcher(1002);
         //剩余空间
-        $users = new Model('users');
+        $db = new Model('users');
         $condition=array('uid'=>$_SESSION['uid']);
         $result=$db->find($condition);
         $used=$result['used']+$filesize;
@@ -362,10 +348,10 @@ class CloudController extends BaseController
         return true;
     }
 
-    //路径是否存在
+    //路径是否存在（传入参数的目录）
     private function is_path_existed($path)
     {
-        if($path[-1]!='/') $path=getName($path)[1];//获取文件所在文件夹
+        if($path[-1]!='/') $path=self::getName($path)[1];//获取文件所在文件夹
         if($path=='/') return true;
         $name=self::getName($path);
         $filename=$name[0];
@@ -373,7 +359,7 @@ class CloudController extends BaseController
         $db=new Model("disk_file");
         $condition=array('uid'=>$_SESSION['uid'],'path'=>$path);
         $result=$db->find($condition);
-        if(empty($result)) return false;
+        if($result==false) return false;
         return true;
 
     }
